자잘한 오류 잡고 수정.

+++ Study/src/main/webapp/WEB-INF/views/study/studyView.jsp
@@ -59,7 +59,10 @@ $(function(){
 	});
 	
 	$("button.removeStudy").click(function(){
-		location.href="deleteStudy.do?sno="+${study.SNO};
+		if(confirm("정말 삭제하시겠습니까?")){
+			location.href="deleteStudy.do?sno="+${study.SNO};
+
+		}
 	});
